Finalmente funciona aquilo do ?nome!!!!!!!!!!!!!!!

+++ Semente/Semente/Models/Apresentacao.cs
@@ -12,7 +12,7 @@ namespace Semente.Models
         public String Forma { get; set; }
         public String Concentracao { get; set; }
         public String Qtd { get; set; }
-        public int MedicamentoId { get; set; }
+        public long MedicamentoId { get; set; }
         public Medicamento Medicamento { get; set; }
         public int FarmacoId { get; set; }
         public Farmaco Farmaco { get; set; }
